pyproject.toml
[tool.poetry]
name = "octopin"
version="0.1.0.dev"
description = "Pin actions and analyse transitive dependencies of GitHub workflows / actions"
authors = [
    "Thomas Neidhart <thomas.neidhart@gmail.com>",
]
license = "MIT License"
readme = "README.md"
homepage = "https://github.com/TinyGearsOrg/octopin"
repository = "https://github.com/TinyGearsOrg/octopin"
keywords = ["python"]

packages = [
    { include = "octopin", from = "src" }
]

include = [
    { path = "*.md", format = "sdist" },
    { path = "src/octopin/py.typed" },
    { path = "docs", format = "sdist" },
    { path = "tests", format = "sdist" }
]

classifiers = [
     'Development Status :: 1 - Alpha',
     'Intended Audience :: Developers',
     'Natural Language :: English',
     'License :: OSI Approved :: MIT License',
     'Programming Language :: Python',
     'Programming Language :: Python :: 3.10',
     'Programming Language :: Python :: 3.11',
     'Typing :: Typed',
]

[tool.poetry.scripts]
octopin = "octopin.console:cli"

[tool.poetry.dependencies]
python               = "^3.10"
cleo                 = "^2.1"
PyYAML               = "^6.0"
semver               = "^3.0"
aiohttp              = "^3.8"
aiohttp-client-cache = "^0.10"
aiosqlite            = "^0.19"
anytree              = "^2.12"

[tool.poetry.group.dev.dependencies]
flake8          = ">=6.0"
pre-commit      = ">=3.3"
black           = ">=23.3"

[tool.poetry.group.test.dependencies]
pytest          = ">=7.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
minversion = "7.3"
testpaths = ["test"]
pythonpath = [".", "src/httpbin"]

[tool.black]
color = true
line-length = 120
target-version = ['py310']
skip-string-normalization = true

[tool.isort]
profile = "black"
line_length = 120
skip_gitignore = true
known_first_party = ['test']

[tool.mypy]
ignore_missing_imports = true
exclude = "dist|build"
